Fix hydration error on navigation away from `series-navigation`
The active route check in the `to` prop resolves to `false` when
navigating away from a page that contains the `series-navigation`
component. In turn the `app-link` component throws an error trying to
resolve `false` as a route with `router.resolve`. The check is unneeded
because of the `v-if` above it.